Rename shared diagnostic and outcome
