+++ packages/alfa-rules/src/common/outcome/contrast.ts
@@ -1,14 +1,14 @@
-import { Contrast } from "../diagnostic/contrast";
+import { Contrast as Diagnostic } from "../diagnostic/contrast";
 import { Err, Ok } from "@siteimprove/alfa-result";
 
-export namespace Outcomes {
+export namespace Contrast {
   export const HasSufficientContrast = (
     highest: number,
     threshold: number,
-    pairings: Array<Contrast.Pairing>
+    pairings: Array<Diagnostic.Pairing>
   ) =>
     Ok.of(
-      Contrast.of(
+      Diagnostic.of(
         `The highest possible contrast of the text is 1:${highest} which is
         above the required contrast of 1:${threshold}`,
         threshold,
@@ -19,10 +19,10 @@ export namespace Outcomes {
   export const HasInsufficientContrast = (
     highest: number,
     threshold: number,
-    pairings: Array<Contrast.Pairing>
+    pairings: Array<Diagnostic.Pairing>
   ) =>
     Err.of(
-      Contrast.of(
+      Diagnostic.of(
         `The highest possible contrast of the text is 1:${highest} which is
         below the required contrast of 1:${threshold}`,
         threshold,
