+++ packages/alfa-rules/test/sia-r66/rule.spec.tsx
@@ -5,8 +5,8 @@ import { RGB, Percentage } from "@siteimprove/alfa-css";
 import { Document, Text } from "@siteimprove/alfa-dom";
 
 import R66 from "../../src/sia-r66/rule";
-import { Contrast } from "../../src/common/diagnostic/contrast";
-import { Outcomes } from "../../src/common/outcome/contrast";
+import { Contrast as Diagnostic } from "../../src/common/diagnostic/contrast";
+import { Contrast as Outcomes } from "../../src/common/outcome/contrast";
 
 import { evaluate } from "../common/evaluate";
 import { passed, failed, cantTell, inapplicable } from "../common/outcome";
@@ -31,7 +31,7 @@ test("evaluate() passes a text node that has sufficient contrast", async (t) => 
   t.deepEqual(await evaluate(R66, { document }), [
     passed(R66, target, {
       1: Outcomes.HasSufficientContrast(21, 7, [
-        Contrast.Pairing.of(rgb(1, 1, 1), rgb(0, 0, 0), 21),
+        Diagnostic.Pairing.of(rgb(1, 1, 1), rgb(0, 0, 0), 21),
       ]),
     }),
   ]);
@@ -55,12 +55,12 @@ test("evaluate() correctly handles semi-transparent backgrounds", async (t) => {
   t.deepEqual(await evaluate(R66, { document }), [
     passed(R66, target1, {
       1: Outcomes.HasSufficientContrast(15.08, 7, [
-        Contrast.Pairing.of(rgb(1, 1, 1), rgb(0.15, 0.15, 0.15), 15.08),
+        Diagnostic.Pairing.of(rgb(1, 1, 1), rgb(0.15, 0.15, 0.15), 15.08),
       ]),
     }),
     failed(R66, target2, {
       1: Outcomes.HasInsufficientContrast(5.74, 7, [
-        Contrast.Pairing.of(rgb(1, 1, 1), rgb(0.4, 0.4, 0.4), 5.74),
+        Diagnostic.Pairing.of(rgb(1, 1, 1), rgb(0.4, 0.4, 0.4), 5.74),
       ]),
     }),
   ]);
@@ -80,12 +80,12 @@ test("evaluate() correctly handles semi-transparent foregrounds", async (t) => {
   t.deepEqual(await evaluate(R66, { document }), [
     passed(R66, target1, {
       1: Outcomes.HasSufficientContrast(14.84, 7, [
-        Contrast.Pairing.of(rgb(0.85, 0.85, 0.85), rgb(0, 0, 0), 14.84),
+        Diagnostic.Pairing.of(rgb(0.85, 0.85, 0.85), rgb(0, 0, 0), 14.84),
       ]),
     }),
     failed(R66, target2, {
       1: Outcomes.HasInsufficientContrast(5.28, 7, [
-        Contrast.Pairing.of(rgb(0.5, 0.5, 0.5), rgb(0, 0, 0), 5.28),
+        Diagnostic.Pairing.of(rgb(0.5, 0.5, 0.5), rgb(0, 0, 0), 5.28),
       ]),
     }),
   ]);
@@ -105,7 +105,7 @@ test("evaluate() passes an 18pt text node with sufficient contrast", async (t) =
   t.deepEqual(await evaluate(R66, { document }), [
     passed(R66, target, {
       1: Outcomes.HasSufficientContrast(5.32, 4.5, [
-        Contrast.Pairing.of(
+        Diagnostic.Pairing.of(
           rgb(0.5019608, 0.5019608, 0.5019608),
           rgb(0, 0, 0),
           5.32
@@ -134,7 +134,7 @@ test("evaluate() passes an 14pt, bold text node with sufficient contrast", async
   t.deepEqual(await evaluate(R66, { document }), [
     passed(R66, target, {
       1: Outcomes.HasSufficientContrast(5.32, 4.5, [
-        Contrast.Pairing.of(
+        Diagnostic.Pairing.of(
           rgb(0.5019608, 0.5019608, 0.5019608),
           rgb(0, 0, 0),
           5.32
@@ -152,7 +152,7 @@ test("evaluate() passes a text node using the user agent default styles", async 
   t.deepEqual(await evaluate(R66, { document }), [
     passed(R66, target, {
       1: Outcomes.HasSufficientContrast(21, 7, [
-        Contrast.Pairing.of(rgb(0, 0, 0), rgb(1, 1, 1), 21),
+        Diagnostic.Pairing.of(rgb(0, 0, 0), rgb(1, 1, 1), 21),
       ]),
     }),
   ]);
@@ -170,7 +170,7 @@ test("evaluate() correctly resolves the `currentcolor` keyword", async (t) => {
   t.deepEqual(await evaluate(R66, { document }), [
     failed(R66, target, {
       1: Outcomes.HasInsufficientContrast(1, 7, [
-        Contrast.Pairing.of(rgb(1, 1, 1), rgb(1, 1, 1), 1),
+        Diagnostic.Pairing.of(rgb(1, 1, 1), rgb(1, 1, 1), 1),
       ]),
     }),
   ]);
@@ -186,7 +186,7 @@ test("evaluate() correctly resolves the `currentcolor` keyword to the user agent
   t.deepEqual(await evaluate(R66, { document }), [
     failed(R66, target, {
       1: Outcomes.HasInsufficientContrast(1, 7, [
-        Contrast.Pairing.of(rgb(0, 0, 0), rgb(0, 0, 0), 1),
+        Diagnostic.Pairing.of(rgb(0, 0, 0), rgb(0, 0, 0), 1),
       ]),
     }),
   ]);
@@ -244,7 +244,7 @@ test("evaluate() passes when a background color with sufficient contrast is inpu
     [
       passed(R66, target, {
         1: Outcomes.HasSufficientContrast(21, 7, [
-          Contrast.Pairing.of(rgb(0, 0, 0), rgb(1, 1, 1), 21),
+          Diagnostic.Pairing.of(rgb(0, 0, 0), rgb(1, 1, 1), 21),
         ]),
       }),
     ]
@@ -271,7 +271,7 @@ test("evaluate() fails when a background color with insufficient contrast is inp
     [
       failed(R66, target, {
         1: Outcomes.HasInsufficientContrast(1, 7, [
-          Contrast.Pairing.of(rgb(0, 0, 0), rgb(0, 0, 0), 1),
+          Diagnostic.Pairing.of(rgb(0, 0, 0), rgb(0, 0, 0), 1),
         ]),
       }),
     ]
@@ -296,8 +296,8 @@ test("evaluate() passes when a linear gradient has sufficient contrast in the be
   t.deepEqual(await evaluate(R66, { document }), [
     passed(R66, target, {
       1: Outcomes.HasSufficientContrast(21, 7, [
-        Contrast.Pairing.of(rgb(0, 0, 0), rgb(1, 1, 1), 21),
-        Contrast.Pairing.of(rgb(0, 0, 0), rgb(0, 0, 0), 1),
+        Diagnostic.Pairing.of(rgb(0, 0, 0), rgb(1, 1, 1), 21),
+        Diagnostic.Pairing.of(rgb(0, 0, 0), rgb(0, 0, 0), 1),
       ]),
     }),
   ]);
@@ -321,8 +321,8 @@ test("evaluate() fails when a linear gradient has insufficient contrast in the b
   t.deepEqual(await evaluate(R66, { document }), [
     failed(R66, target, {
       1: Outcomes.HasInsufficientContrast(1, 7, [
-        Contrast.Pairing.of(rgb(0, 0, 0), rgb(0, 0, 0), 1),
-        Contrast.Pairing.of(rgb(0, 0, 0), rgb(0, 0, 0), 1),
+        Diagnostic.Pairing.of(rgb(0, 0, 0), rgb(0, 0, 0), 1),
+        Diagnostic.Pairing.of(rgb(0, 0, 0), rgb(0, 0, 0), 1),
       ]),
     }),
   ]);
@@ -346,7 +346,7 @@ test(`evaluate() correctly merges semi-transparent background layers against a
   t.deepEqual(await evaluate(R66, { document }), [
     passed(R66, target, {
       1: Outcomes.HasSufficientContrast(10.41, 7, [
-        Contrast.Pairing.of(rgb(1, 1, 1), rgb(0.25, 0.25, 0.25), 10.41),
+        Diagnostic.Pairing.of(rgb(1, 1, 1), rgb(0.25, 0.25, 0.25), 10.41),
       ]),
     }),
   ]);
